release 3.1.180 source code

+++ services/hss/src/main/java/com/huaweicloud/sdk/hss/v5/model/ShowRaspProtectStatisticsResponse.java
@@ -19,7 +19,7 @@ public class ShowRaspProtectStatisticsResponse extends SdkResponse {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "anti_tampering_num")
 
-    private Long antiTamperingNum;
+    private Integer antiTamperingNum;
 
     public ShowRaspProtectStatisticsResponse withProtectHostNum(Long protectHostNum) {
         this.protectHostNum = protectHostNum;
@@ -27,7 +27,7 @@ public class ShowRaspProtectStatisticsResponse extends SdkResponse {
     }
 
     /**
-     * 防护主机数
+     * **参数解释** 当前项目（或指定企业项目）下已添加应用防护的云服务器总数，统计范围为所有已启用RASP防护的主机 **取值范围** 取值0-为该项目下云服务器总数量（无上限，实际受账号资源配额限制） 
      * minimum: 0
      * maximum: 9223372036854775807
      * @return protectHostNum
@@ -40,22 +40,22 @@ public class ShowRaspProtectStatisticsResponse extends SdkResponse {
         this.protectHostNum = protectHostNum;
     }
 
-    public ShowRaspProtectStatisticsResponse withAntiTamperingNum(Long antiTamperingNum) {
+    public ShowRaspProtectStatisticsResponse withAntiTamperingNum(Integer antiTamperingNum) {
         this.antiTamperingNum = antiTamperingNum;
         return this;
     }
 
     /**
-     * 防御篡改攻击数
+     * **参数解释** 近7天内当前项目（或指定企业项目）下RASP防护成功拦截的篡改类攻击总数，与功能介绍中'近七天微服务RASP攻击数量'对应 **取值范围** 取值0-无上限（实际受攻击频次限制） 
      * minimum: 0
      * maximum: 9223372036854775807
      * @return antiTamperingNum
      */
-    public Long getAntiTamperingNum() {
+    public Integer getAntiTamperingNum() {
         return antiTamperingNum;
     }
 
-    public void setAntiTamperingNum(Long antiTamperingNum) {
+    public void setAntiTamperingNum(Integer antiTamperingNum) {
         this.antiTamperingNum = antiTamperingNum;
     }
