+++ services/swr/src/main/java/com/huaweicloud/sdk/swr/v2/model/Artifact.java
@@ -66,7 +66,7 @@ public class Artifact {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "size")
 
-    private Integer size;
+    private Long size;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "tags")
@@ -243,7 +243,7 @@ public class Artifact {
         this.pushTime = pushTime;
     }
 
-    public Artifact withSize(Integer size) {
+    public Artifact withSize(Long size) {
         this.size = size;
         return this;
     }
@@ -252,11 +252,11 @@ public class Artifact {
      * 制品大小，单位：Byte
      * @return size
      */
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(Long size) {
         this.size = size;
     }
